// models/team.enum.ts
export enum Team {
  Vimpeli = 'Vimpeli',
  Sotkamo = 'Sotkamo',
  Alajarvi = 'Alajärvi',
  Kitee = 'Kitee',
  Tampere = 'Tampere',
  Hyvinkaa = 'Hyvinkää',
  Kempele = 'Kempele',
  Imatra = 'Imatra',
  Joensuu = 'Joensuu',
  Kouvola = 'Kouvola',
  Seinajoki = 'Seinäjoki',
  Koskenkorva = 'Kossu',
  Pattijoki = 'Pattijoki',
  Vimpeli_2 = 'Vimpeli',
  Sotkamo_2 = 'Sotkamo',
  Alajarvi_2 = 'Alajärvi',
  Kitee_2 = 'Kitee',
  Tampere_2 = 'Tampere',
  Hyvinkaa_2 = 'Hyvinkää',
  Kempele_2 = 'Kempele',
  Imatra_2 = 'Imatra',
  Joensuu_2 = 'Joensuu',
  Kouvola_2 = 'Kouvola',
  Seinajoki_2 = 'Seinäjoki',
  Koskenkorva_2 = 'Kossu',
  Pattijoki_2 = 'Pattijoki',
  
};